Ignore killall alias when killall already exists

+++ shared/etc/initrc.sh
@@ -134,7 +134,10 @@ withtimeout() {
     )
 }
 
-echoAlias killall 'test -n "$1" && ps -A | grep "$1" | cut -d" " -f1 | xargs kill'
+if ! test -x killall ; then
+    echoAlias killall 'test -n "$1" && ps -A | grep "$1" | cut -d" " -f1 | xargs kill'
+fi
+
 echoAlias npm-unlink 'npm rm --global $*'
 echoAlias find-symlinks 'find -L . -xtype l -ls'
 echoAlias reload '. ~/.bashrc'
